implement sbrk, brk, fix malloc, and finish serial transfer

+++ binutils/lnk/io.c
@@ -49,21 +49,21 @@ void raw_output_word(uint16_t word){
 /* functions for managing the layout of the different segments, and the data in them */
 
 /* the size of each segment (in bytes, not counting obj meta bytes) */
-uint16_t seg_size[4];
+uint32_t seg_size[4];
 /* the start of each segment in memory (in bytes) */
-uint16_t seg_start[4];
+uint32_t seg_start[4];
 
 /* the offset of each file's segments */
-uint16_t in_segs_start[MAX_FILES][4];
+uint32_t in_segs_start[MAX_FILES][4];
 
 /* current seg being written to */
 uint8_t cur_seg;
 /* current write location in each seg (offset in seg) */
-uint16_t seg_pos[4];
+uint32_t seg_pos[4];
 
 /* start of each file in symbol tables (only for obj output) */
-uint16_t defined_start[MAX_FILES];
-uint16_t extern_start[MAX_FILES];
+uint32_t defined_start[MAX_FILES];
+uint32_t extern_start[MAX_FILES];
 
 /* create the segment layout, and write the header if requested
    the header contains the start page (5 bits) and number of pages (5 bits) for each segment
@@ -71,7 +71,7 @@ uint16_t extern_start[MAX_FILES];
    if the header is created, it should be loaded into memory with the rest of the program - scplnk accounts for the offset
    do_pages is whether to arrange segments on page boundries or not. If not, they are arranged on word boundries */
 void bin_create_segs(uint8_t do_head, uint8_t do_pages){
-    uint16_t offset = 0;
+    uint32_t offset = 0;
 
     /* make space for header */
     if(do_head){
@@ -83,7 +83,7 @@ void bin_create_segs(uint8_t do_head, uint8_t do_pages){
         /* offset += seg_size[s]; */
         for(int i = 0; in_objs[i].file; i++){
             /* divide by two b/c sizes in header includes meta-bytes */
-            uint16_t size = in_objs[i].segs.segs[s].size / 2;
+            uint32_t size = in_objs[i].segs.segs[s].size / 2;
             /* sum total seg_size */
             seg_size[s] += size;
             /* set in file specific segment starts (needed for extern lookups) */
@@ -133,7 +133,7 @@ void obj_out_create_segs(){
 
         for(int i = 0; in_objs[i].file; i++){
             /* divide by two b/c sizes in header includes meta-bytes */
-            uint16_t size = in_objs[i].segs.segs[s].size;
+            uint32_t size = in_objs[i].segs.segs[s].size;
             /* sum total seg_size */
             seg_size[s] += size;
             /* set in file specific segment starts (needed for extern lookups) */
